requirements.txt
fastapi
cryptography
passlib
uvicorn
git+https://github.com/jtgrassie/pyrx
tinyec